% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.h.R
\name{pareto}
\alias{pareto}
\title{Pareto Chart}
\usage{
pareto(data, x, counts = NULL, angle = 0)
}
\arguments{
\item{data}{the data as a data frame}

\item{x}{a string naming the variable from \code{data} that contains the
values used for the chart}

\item{counts}{a string naming the variable from \code{data} that contains
the  counts for the values (optional)}

\item{angle}{a number from 0 to 45 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$pareto} \tab \tab \tab \tab \tab a Pareto chart \cr
}
}
\description{
Function for making pareto charts.
}
\examples{
set.seed(1337)

X <- sample(c('A','B','C','D','E','F'), 100, replace=TRUE)
dat <- data.frame(X = X)

pareto(dat, x = 'X')

}
