% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatrr-package.R
\docType{package}
\name{scatrr-package}
\alias{scatrr-package}
\alias{scatrr}
\title{scta client}
\description{
scta client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
