% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scta_search.R
\name{scta_search}
\alias{scta_search}
\title{Search}
\usage{
scta_search(q = NULL, parse = TRUE, ...)
}
\arguments{
\item{q}{query terms}

\item{parse}{to parse to data.fram'es if possible or to list}

\item{...}{curl options}
}
\description{
Search
}
\examples{
\dontrun{
res <- scta_search(q = 'potentia')
res$resources
res$within
res$`next`
res$`@context`
}
}
